Refactor FlowUseCase
